pyproject.toml
[tool.poetry]
name = "12-17-quiz"
version = "0.1.0"
description = ""
authors = ["KK-Huang86 <iankk0806@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
